[VB-26] general refactoring, adding CalendarComponent
