+++ VolleyBolley/Modules/Home/View/HomeViewController.swift
@@ -5,6 +5,7 @@
 //  Created by Nikolai Eremenko
 //
 
+import Combine
 import UIKit
 
 protocol HomeViewProtocol: AnyObject {
@@ -13,9 +14,17 @@ protocol HomeViewProtocol: AnyObject {
     func displayError(message: String)
 }
 
+extension HomeViewController: CalendarComponentDelegate {
+	func didSelectDate(_ date: Date) {
+		updateDateLabel(with: date)
+	}
+}
+
 final class HomeViewController: BaseViewController, HomeViewProtocol {
 
     // MARK: - Private Properties
+
+	private lazy var calendarComponent: CalendarComponentProtocol = CalendarComponent(delegate: self)
 
     private let presenter: HomePresenterProtocol
 
@@ -74,6 +83,8 @@ final class HomeViewController: BaseViewController, HomeViewProtocol {
     private func setupLayout() {
         setupConstraintsNavBar()
         setupConstraintsLabel()
+
+		setupCalendar()
     }
 
     // MARK: - Constraints
@@ -97,6 +108,28 @@ final class HomeViewController: BaseViewController, HomeViewProtocol {
             label.centerYAnchor.constraint(equalTo: view.centerYAnchor)
         ])
     }
+
+	private func setupCalendar() {
+		let calendarVC = calendarComponent.createCalendarViewController()
+		addChild(calendarVC)
+		view.addSubviews(calendarVC.view)
+		calendarVC.didMove(toParent: self)
+
+		NSLayoutConstraint.activate([
+			calendarVC.view.topAnchor.constraint(equalTo: label.bottomAnchor, constant: 10),
+			calendarVC.view.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
+			calendarVC.view.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
+			calendarVC.view.widthAnchor.constraint(equalToConstant: 319),
+			calendarVC.view.heightAnchor.constraint(equalToConstant: 330)
+		])
+	}
+
+	private func updateDateLabel(with date: Date) {
+		let formatter = DateFormatter()
+		formatter.dateStyle = .full
+		formatter.timeStyle = .none
+		label.text = "Выбранная дата: \(formatter.string(from: date))"
+	}
 }
 
 // MARK: - Preview
